Fix compilation issues on Visual Studio 2008. Also added precompiled headers, it needs to be disabled explicitly for MOC files, creation of the precompiled headers is executed from sc_main.cpp and main.cpp.
git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@11246 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/class_modules/sc_priest.cpp
@@ -1246,7 +1246,7 @@ void priest_spell_t::trigger_shadowy_apparition( priest_t* p )
 {
   if ( !p -> shadowy_apparition_free_list.empty() )
   {
-    for ( int i = p-> resources.current[ RESOURCE_SHADOW_ORB ]; i > 0; i-- )
+    for ( int i = (int)p -> resources.current[ RESOURCE_SHADOW_ORB ]; i > 0; i-- )
     {
       spell_t* s = p -> shadowy_apparition_free_list.front();
